Cyclomatic Complexity measured with Lizard

+++ testing/measures/CyclomaticComplexityBaseMeasure.py
@@ -3,7 +3,7 @@ from github.Repository import Repository
 from domain.model.Measure import BaseMeasure
 
 
-class CyclomaticComplexityBaseMeasure(BaseMeasure[int]):
+class CyclomaticComplexityBaseMeasure(BaseMeasure[float]):
     def __init__(self, visitor=None):
         super().__init__("Cyclomatic Complexity Base Measure", "", 1, visitor)
